refactor: 기존페이지 분리
